Follow-up to #2896878 by drunken monkey: Fixed tests on Drupal 8.3.

+++ tests/src/Functional/ViewsTest.php
@@ -389,7 +389,9 @@ class ViewsTest extends SearchApiBrowserTestBase {
       // used (so we could reproduce the bug if it's there).
       // @todo Remove "cache.render" once we depend on Drupal 8.4+.
       \Drupal::getContainer()->get('cache.render')->deleteAll();
-      \Drupal::getContainer()->get('cache.page')->deleteAll();
+      if (\Drupal::getContainer()->has('cache.page')) {
+        \Drupal::getContainer()->get('cache.page')->deleteAll();
+      }
       \Drupal::getContainer()->get('cache.dynamic_page_cache')->deleteAll();
       $this->submitForm([], 'Search');
       $this->assertSession()->addressEquals('search-api-test');
